back-feat: add classe abstrata para validação de conta post e put
